web/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
torch==2.5.1+cu124
torchvision==0.20.1+cu124
numpy==1.26.3
matplotlib==3.8.2
scikit-image==0.25.0
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
pillow==10.2.0
python-multipart==0.0.19